ai-brain/requirements.txt
llama-cpp-python==0.2.90
huggingface-hub>=0.20.0
